move examples to root. index.ts as scaffold
