+++ examples/FullExample.ts
@@ -1,15 +1,17 @@
-import fetch from 'node-fetch'
-
-import { FlatfileRecord, FlatfileRecords } from '@flatfile/hooks'
 import {
-  Sheet,
-  Workbook,
-  TextField,
   BooleanField,
+  DateField,
+  Message,
   NumberField,
   OptionField,
-  Message,
+  Portal,
+  Sheet,
+  TextField,
+  Workbook,
 } from '@flatfile/configure'
+
+import { FlatfileRecord, FlatfileRecords } from '@flatfile/hooks'
+import fetch from 'node-fetch'
 
 const Employees = new Sheet(
   'Employees',
@@ -18,14 +20,18 @@ const Employees = new Sheet(
       required: true,
       description: 'Given name',
     }),
-    lastName: TextField(),
+    lastName: TextField({
+      compute: (v: any) => {
+        return `Rock`
+      },
+    }),
     fullName: TextField(),
 
     stillEmployed: BooleanField(),
     department: OptionField({
       label: 'Department',
       options: {
-        engineering: 'Engineering',
+        engineering: { label: 'Engineering' },
         hr: 'People Ops',
         sales: 'Revenue',
       },
@@ -48,13 +54,14 @@ const Employees = new Sheet(
         }
       },
     }),
+    startDate: DateField()
   },
   {
     allowCustomFields: true,
     readOnly: true,
     recordCompute: (record) => {
       const fullName = `{record.get('firstName')} {record.get('lastName')}`
-      record.set('fullhName', fullName)
+      record.set('fullName', fullName)
       return record
     },
     batchRecordsCompute: async (payload: FlatfileRecords<any>) => {
@@ -66,16 +73,22 @@ const Employees = new Sheet(
       })
       const result = await response.json()
       payload.records.map(async (record: FlatfileRecord) => {
-        await record.set('fromHttp', result.info.postgres.status)
+        record.set('fromHttp', result.info.postgres.status)
       })
     },
   }
 )
 
+const EmployeesPortal = new Portal({
+  name: 'EmployeesPortal',
+  sheet: 'Employees'
+})
+
 export default new Workbook({
-  name: 'Migration stage1',
-  namespace: 'MyCompany',
+  name: 'Employees',
+  namespace: 'employee',
   sheets: {
     Employees,
   },
+  portals: [EmployeesPortal],
 })
